Refactor duplicated fixtures introduced in #59
