+++ tests/conftest.py
@@ -10,6 +10,6 @@ def _to_module_string(path: str) -> str:
 
 pytest_plugins = [
     _to_module_string(fixture.relative_to(_THIS_DIR.parent).as_posix())
-    for fixture in (Path(f"{_THIS_DIR}/fixtures").glob("*.py"))
+    for fixture in (Path(f"{_THIS_DIR}/fixtures").rglob("*.py"))
     if "__init__" not in str(fixture)
 ]
