+++ tests/test_graph/test_model_constructor.py
@@ -1,43 +1,14 @@
 import re
 from collections.abc import Callable
-from typing import Any, Concatenate, TypeAlias
+from typing import Any
 
-import jax
 import numpy as np
 import numpy.typing as npt
 import numpyro
 import pytest
 from numpyro.distributions import Normal
-from numpyro.infer import MCMC, NUTS
 
 from causalprog.graph import DistributionNode, Graph, ParameterNode
-
-# TODO: Refactor into conftest to share with test_node/test_create_model_site.py.
-# Similarly for any other graphs here that are shared... and decorator-fixtures
-MCMCRunner: TypeAlias = Callable[Concatenate[Callable, ...], MCMC]
-
-
-@pytest.fixture(scope="session")
-def mcmc_default_options() -> dict[str, float]:
-    return {"num_warmup": 500, "num_samples": 1000}
-
-
-@pytest.fixture
-def run_nuts_mcmc(
-    rng_key: jax.Array,
-) -> MCMCRunner:
-    def inner(model, *, nuts_kwargs=None, mcmc_kwargs=None) -> MCMC:
-        if not nuts_kwargs:
-            nuts_kwargs = {}
-        if not mcmc_kwargs:
-            mcmc_kwargs = {}
-
-        kernel = NUTS(model, **nuts_kwargs)
-        mcmc = MCMC(kernel, **mcmc_kwargs)
-        mcmc.run(rng_key)
-        return mcmc
-
-    return inner
 
 
 @pytest.fixture
@@ -84,7 +55,7 @@ def two_normal_graph_expected_model() -> Callable[..., Callable[[], None]]:
 def test_model_constructor(
     two_normal_graph: Graph,
     two_normal_graph_expected_model: Callable[[float, float], Callable[[], None]],
-    run_nuts_mcmc: MCMCRunner,
+    run_nuts_mcmc,
     mcmc_default_options: dict[str, Any],
     collection_of_parameter_values: tuple[dict[str, npt.ArrayLike], ...] = (
         {"mu_x": 0.0, "nu_y": 1.0},
